+++ js/server.js
@@ -63,7 +63,7 @@ function Server(config, callback) {
 
 	app.use("/js", express.static(__dirname));
 
-	const directories = ["/config", "/css", "/fonts", "/modules", "/vendor", "/translations", "/tests/configs"];
+	const directories = ["/config", "/css", "/fonts", "/modules", "/" + config.paths.modules, "/vendor", "/translations", "/tests/configs"];
 	for (const directory of directories) {
 		app.use(directory, express.static(path.resolve(global.root_path + directory)));
 	}